+++ src/client/Vo/selloutVO.java
@@ -3,26 +3,26 @@ package client.Vo;
 import java.io.Serializable;
 
 public class selloutVO implements Serializable{
-	private Long kind;
+	private Double kind;
 	private String keyno;
 	private String note;
 	private String oper;
-	private Long ischeck;
-	private Long isred;
+	private Double ischeck;
+	private Double isred;
 	private String consumer;
 	private String server;
 	private String base;
 	private String goodsoutlist;
-	private Long sumall;
-	private Long cut;
-	private Long voucher;
-	private Long finalsum;
+	private Double sumall;
+	private Double cut;
+	private Double voucher;
+	private Double finalsum;
 
-	public Long getKind() {
+	public Double getKind() {
 		return kind;
 	}
 
-	public void setKind(Long kind) {
+	public void setKind(Double kind) {
 		this.kind = kind;
 	}
 
@@ -50,19 +50,19 @@ public class selloutVO implements Serializable{
 		this.oper = oper;
 	}
 
-	public Long getIscheck() {
+	public Double getIscheck() {
 		return ischeck;
 	}
 
-	public void setIscheck(Long ischeck) {
+	public void setIscheck(Double ischeck) {
 		this.ischeck = ischeck;
 	}
 
-	public Long getIsred() {
+	public Double getIsred() {
 		return isred;
 	}
 
-	public void setIsred(Long isred) {
+	public void setIsred(Double isred) {
 		this.isred = isred;
 	}
 
@@ -98,39 +98,39 @@ public class selloutVO implements Serializable{
 		this.goodsoutlist = goodsoutlist;
 	}
 
-	public Long getSumall() {
+	public Double getSumall() {
 		return sumall;
 	}
 
-	public void setSumall(Long sumall) {
+	public void setSumall(Double sumall) {
 		this.sumall = sumall;
 	}
 
-	public Long getCut() {
+	public Double getCut() {
 		return cut;
 	}
 
-	public void setCut(Long cut) {
+	public void setCut(Double cut) {
 		this.cut = cut;
 	}
 
-	public Long getVoucher() {
+	public Double getVoucher() {
 		return voucher;
 	}
 
-	public void setVoucher(Long voucher) {
+	public void setVoucher(Double voucher) {
 		this.voucher = voucher;
 	}
 
-	public Long getFinalsum() {
+	public Double getFinalsum() {
 		return finalsum;
 	}
 
-	public void setFinalsum(Long finalsum) {
+	public void setFinalsum(Double finalsum) {
 		this.finalsum = finalsum;
 	}
 
-	public selloutVO(Long kind, String keyno, String note, String oper, Long ischeck, Long isred, String consumer, String server, String base, String goodsoutlist, Long sumall, Long cut, Long voucher, Long finalsum) {
+	public selloutVO(Double kind, String keyno, String note, String oper, Double ischeck, Double isred, String consumer, String server, String base, String goodsoutlist, Double sumall, Double cut, Double voucher, Double finalsum) {
 
 		this.kind = kind;
 		this.keyno = keyno;
