+++ src/server/Po/packPO.java
@@ -5,7 +5,7 @@ public class packPO implements Serializable{
 	private String keyno;
 	private String goodsno;
 	private String goodsname;
-	private Long price;
+	private Double price;
 
     public packPO() {
 
@@ -35,15 +35,15 @@ public class packPO implements Serializable{
 		this.goodsname = goodsname;
 	}
 
-	public Long getPrice() {
+	public Double getPrice() {
 		return price;
 	}
 
-	public void setPrice(Long price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
 
-	public packPO(String keyno, String goodsno, String goodsname, Long price) {
+	public packPO(String keyno, String goodsno, String goodsname, Double price) {
 
 		this.keyno = keyno;
 		this.goodsno = goodsno;
