+++ src/server/Po/stockexceptionPO.java
@@ -2,26 +2,26 @@ package server.Po;
 import java.io.Serializable;
 
 public class stockexceptionPO implements Serializable {
-    private Long kind;
+    private Double kind;
     private String note;
     private String oper;
-    private Long ischeck;
+    private Double ischeck;
     private String isred;
     private String goodsno;
     private String goodsname;
-    private Long numinbase;
-    private Long numinsys;
+    private Double numinbase;
+    private Double numinsys;
     private String keyno;
 
     public stockexceptionPO() {
 
     }
 
-    public Long getKind() {
+    public Double getKind() {
         return kind;
     }
 
-    public void setKind(Long kind) {
+    public void setKind(Double kind) {
         this.kind = kind;
     }
 
@@ -41,11 +41,11 @@ public class stockexceptionPO implements Serializable {
         this.oper = oper;
     }
 
-    public Long getIscheck() {
+    public Double getIscheck() {
         return ischeck;
     }
 
-    public void setIscheck(Long ischeck) {
+    public void setIscheck(Double ischeck) {
         this.ischeck = ischeck;
     }
 
@@ -73,19 +73,19 @@ public class stockexceptionPO implements Serializable {
         this.goodsname = goodsname;
     }
 
-    public Long getNuminbase() {
+    public Double getNuminbase() {
         return numinbase;
     }
 
-    public void setNuminbase(Long numinbase) {
+    public void setNuminbase(Double numinbase) {
         this.numinbase = numinbase;
     }
 
-    public Long getNuminsys() {
+    public Double getNuminsys() {
         return numinsys;
     }
 
-    public void setNuminsys(Long numinsys) {
+    public void setNuminsys(Double numinsys) {
         this.numinsys = numinsys;
     }
 
@@ -97,7 +97,7 @@ public class stockexceptionPO implements Serializable {
         this.keyno = keyno;
     }
 
-    public stockexceptionPO(Long kind, String note, String oper, Long ischeck, String isred, String goodsno, String goodsname, Long numinbase, Long numinsys, String keyno) {
+    public stockexceptionPO(Double kind, String note, String oper, Double ischeck, String isred, String goodsno, String goodsname, Double numinbase, Double numinsys, String keyno) {
 
         this.kind = kind;
         this.note = note;
